Add server variable condition mangling
Add mangling of SERVER_VARIABLE_(PRESENT|ABSENT) conditions such that
var_value="1" is added to conditions that have any_value but not a
var_value.  This is a workaround for a bug in a0.8.19 where the
any_value condition doesn't work without var_value set.

Add similar mangling where the condition type is swapped to the logical
opposite for these condition where neither var_value nor any_value is
set.  This is due to highly unintuitive behaviour where the test is
effectively the reverse of what would logicaly make sense given the
name.

Add documentation for these behavoirs and swtiches in the diagen_option
to disable these behaviours.